// wrench-works-web/src/api/generated/models/bookingDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Wrench Works API
 * Workshop management SaaS API
 * OpenAPI spec version: v1
 */
import type { BookingDtoZoneColor } from './bookingDtoZoneColor';
import type { BookingDtoVehicleDisplay } from './bookingDtoVehicleDisplay';
import type { BookingDtoNotes } from './bookingDtoNotes';
import type { BookingDtoJobId } from './bookingDtoJobId';

export interface BookingDto {
  id: string;
  zoneId: string;
  zoneName: string;
  zoneColor: BookingDtoZoneColor;
  customerId: string;
  customerName: string;
  vehicleId: string;
  vehicleDisplay: BookingDtoVehicleDisplay;
  title: string;
  startUtc: string;
  endUtc: string;
  notes: BookingDtoNotes;
  status: string;
  jobId: BookingDtoJobId;
  createdAtUtc: string;
}